NeighborhoodApproximationIndex

NeighborhoodApproximationIndex.KnrIndex

NeighborhoodApproximationIndex.references

SimilaritySearch.optimize!

SimilaritySearch.search

NeighborhoodApproximationIndex.KnrIndex — Method

KnrIndex(
    dist::SemiMetric,
    db;
    numcenters=ceil(Int, sqrt(length(db))),
    refs=:rand,
    centers=:delone,
    k=5,
    pools=nothing,
    parallel_block=Threads.nthreads() > 1 ? 1024 : 1,
    verbose=true
)

Creates a KnrIndex, high level interface.

dist: Distance object (a SemiMetric object, see Distances.jl)

db: Objects to be indexed Please note that these indexes will be created with generic parameters

refs: the se of reference, only used if centers=nothing

pools: an object with preallocated caches specific for KnrIndex, if pools=nothing it will use default caches.

parallel_block Parallel construction works on batches, this is the size of these blocks

verbose true if you want to see messages

NeighborhoodApproximationIndex.references — Method

references(dist::SemiMetric, db; <kwargs>) -> SubDatabase

Computes a set of references, a sample of db, using the KCenters specification, it is a wrapper to kcenters function in the KCenters.jl package. The set of references are computed taking into account the specified r but also the distance function dist.

Arguments

dist: a distance function

db: the database to be sampled

Keyword arguments

k::Int: the number of centers to compute, defaults to sqrt(|db|)

maxiters::Int: number of iterationso of the Lloyd algorithm that should be applied on the initial computation of centers, that is, maxiters > 0 applies maxiters iterations of the algorithm.

tol::Float64: change tolerance to stop the Lloyd algorithm (error changes smaller than tol among iterations will stop the algorithm)

initial: initial centers or a strategy to compute initial centers, symbols :rand, :fft, and :dnet.

There are several interactions between initial values and parameter interactions (described in KCenters object), for instance, the maxiters > 0 apply the Lloyd's algorithm to the initial computation of references.

if initial=:rand:

maxiters = 0 will retrieve a simple random sampling

maxiters > 0' achieve kmeans-centroids,maxiters` should be set appropiately for the the dataset

if initial=:dnet:

maxiters = 0 computes a pure density-net

maxiters > 0 will compute a kmeans centroids but with an initialization based on the dnet

if initial=:fft:

maxiters = 0 computes k centers with the farthest first traversal algorithm

maxiters > 0 will use the FFT based kcenters as initial points for the Lloyd algorithm

Note 1: maxiters > 0 needs to compute centroids and these centroids should be defined for the specific data model, and also be of use in the specific metric distance and error function.

Note 2: The error function is defined as the mean of distances of all objects in db to its associated nearest centers in each iteration.

Note 3: The computation of references on large databases can be prohibitive, in these cases please consider to work on a sample of db

SimilaritySearch.optimize! — Function

optimize!(
    index::KnrIndex,
    kind::ErrorFunction=ParetoRecall();
    queries=nothing,
    queries_ksearch=10,
    queries_size=64,
    initialpopulation=8,
    verbose=false,
    space=KnrOptSpace(),
    params=SearchParams(; maxpopulation=8, bsize=4, mutbsize=8, crossbsize=2, tol=-1.0, maxiters=8, verbose),
)

Tries to configure the index to achieve the specified performance (kind). The optimization procedure is an stochastic search over the configuration space yielded by kind and queries.

Arguments

index: the KnrIndex to be optimized

kind: The kind of optimization to apply, it can be ParetoRecall(), ParetoRadius() or MinRecall(r) where r is the expected recall (0-1, 1 being the best quality but at cost of the search time)

Keyword arguments

queries: the set of queries to be used to measure performances, a validation set. It can be an AbstractDatabase or nothing.

queries_ksearch: the number of neighbors to retrieve for queries

queries_size: if queries===nothing then a sample of the already indexed database is used, queries_size is the size of the sample.

initialpopulation: the initial sample for the optimization procedure

space: defines the search space

params: the parameters of the solver, see search_models function from SearchModels.jl package for more information.

verbose: controls if the procedure is verbose or not

SimilaritySearch.search — Method

search(idx::KnrIndex, q, res::KnnResult; ksearch=idx.ksearch, rerank=idx.rerank, pools=getpools(D))

Searches nearest neighbors of q inside the index under the distance function dist.